remove(cgo_harness): Remove dead code and gate C-bench helpers
- Move C-benchmark helper functions to a new treesitter_c_bench-tagged file so they only compile when CGO C benchmarks are enabled
- Remove unused prepareEditedBenchmarkSource helper from benchmark_helpers_test.go
- Remove unused parseRealCorpusGoIncremental and parseRealCorpusGoIncrementalWithPhase wrappers from real-corpus parity tests
- Remove unused tryParseCompleteRealCorpusCFull helper from real-corpus parity tests
- Remove unused collectCandidatesWithNamesFromRoot from build_real_corpus command
- Remove unused insertSpaceAt helper from parity_cgo_test.go
- Remove stray break after unconditional assignment in safeEditOffset
